Quicksort algorithm parallelized with OpenMP
Achieved an approximate 30% speedup of the Quicksort algorithm on large array inputs using OpenMP.

Idea is that quicksort breaks up the array into two chunks to be sorted after initial sort. So put these two chunks in their own threads and let them sort themselves. 

Adjusted base case to be larger so that we reduce number of threads we are creating. A good number for an array of size 200,000 was 5,000 to achieve an approximate 30% speedup.